requirements.txt
streamlit>=1.20.0
tensorflow>=2.11.0
Pillow
numpy
matplotlib
opencv-python-headless
